Make video blocks have same "pending event placeholder" as on video page
The logic was just moved to the `Player` component. Also, the props of
`Player` were somewhat simplified by grouping most into one `event`
prop that can be passed more easily.
